api/requirements.txt
asttokens==3.0.0
blinker==1.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
exceptiongroup==1.3.0
executing==2.2.0
Faker==37.4.0
Flask==3.1.1
flask-cors==6.0.1
Flask-SQLAlchemy==3.1.1
fonttools==4.58.4
greenlet==3.2.3
idna==3.10
ipykernel==6.29.5
ipython==8.37.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.2.6
packaging==25.0
pandas==2.3.0
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pyzmq==27.0.0
requests==2.32.4
scikit-learn==1.7.0
scipy==1.15.3
six==1.17.0
SQLAlchemy==2.0.41
stack-data==0.6.3
threadpoolctl==3.6.0
tornado==6.5.1
traitlets==5.14.3
typing_extensions==4.14.0
tzdata==2025.2
Unidecode==1.4.0
urllib3==2.5.0
wcwidth==0.2.13
Werkzeug==3.1.3
pytest
pytest-cov
